pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zenith-card-processor"
version = "1.0.0"
description = "Configurable card sprite sheet processor for creating printable collages and PDFs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "Pillow>=10.0.0",
    "pypdf>=3.0.0",
    "reportlab>=4.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
